+++ app/src/main/java/ru/practicum/android/diploma/core/data/mapper/FilterIndustryMapper.kt
@@ -1,0 +1,11 @@
+package ru.practicum.android.diploma.core.data.mapper
+
+import ru.practicum.android.diploma.core.data.network.dto.FilterIndustryDto
+import ru.practicum.android.diploma.core.domain.model.FilterIndustry
+
+fun FilterIndustryDto.toDomain(): FilterIndustry {
+    return FilterIndustry(
+        id = id,
+        name = name
+    )
+}
