+++ app/src/main/java/ru/practicum/android/diploma/core/data/network/dto/FilterAreaDto.kt
@@ -7,12 +7,5 @@ data class FilterAreaDto(
     val name: String,
     val parentId: Int?,
     val areas: List<FilterAreaDto>
-) {
-    fun toDomain(): FilterArea {
-        return FilterArea(
-            id = id,
-            name = name
-        )
-    }
-}
+)
