+++ app/src/main/java/ru/practicum/android/diploma/core/data/network/dto/FilterIndustryDto.kt
@@ -5,11 +5,4 @@ import ru.practicum.android.diploma.core.domain.model.FilterIndustry
 data class FilterIndustryDto(
     val id: Int,
     val name: String
-) {
-    fun FilterIndustryDto.toDomain(): FilterIndustry {
-        return FilterIndustry(
-            id = id,
-            name = name
-        )
-    }
-}
+)
